Added Register, Logout, and Login functionality.
Added Register, Logout, and Login functionality.

You can get to the Login page by clicking on the Avatar in the navbar.
You can get to the Register page through the Login page. You can logout
by clicking on the Avatar while logged in.

This navigation is only temporary. We might have to change the style of
the navbar so we can add a Login/Logout button.