Login Front-end Update + Custom Panel
Added a custom panel style called panel-eph. You can use this with a
panel in place of the provided classes (panel-default, panel-danger,
etc) and it will generate a transparent panel compatable with the site.

Updated the look of the login which makes use of the custom panel.
Elements are organized and responsive for mobile and desktop.